+++ src/Filter/Type/CheckboxType.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * Copyright (c) 2022 Heimrich & Hannot GmbH
+ * Copyright (c) 2023 Heimrich & Hannot GmbH
  *
  * @license LGPL-3.0-or-later
  */
@@ -80,5 +80,10 @@ class CheckboxType extends AbstractType
         }
 
         return $options;
+    }
+
+    public static function normalizeValue($value)
+    {
+        return (bool) $value ? '1' : '';
     }
 }